arm64: Split EFI map parsing code to a common file
Motivation: I would like to enable the EFI memory map parsing on riscv,
which is preferable to parsing memory regions from the device tree. I
could easily duplicate the functionality, but it can be shared without
much extra effort. There are a couple of similar implementations
existing already.

In this commit, the arm64 version of the code is moved to the new file,
and enabled for this platform. No functional change intended. armv7 and
riscv will follow.

amd64 could be converted to use this interface too, but it doesn't use
the common "physmem" interfaces for managing physical memory, and would
therefore require further tweaks. I am not inclined to make this change
at this time.

Reviewed by:	imp, andrew
Discussed with:	emaste (copyright attribution)
Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D47746
